requirements.txt
altair
numpy
pandas
pydeck
streamlit
langchain
langchain_experimental
tabulate
google-cloud-aiplatform
os
json
google.oauth2
google.cloud